Split Project page into separate components
